overdue commit for logs 13/14: 13 - add to string and nicer ptree printing 13 - add edge case checks for empty source files 14 - setup helper scripts for environment 14 - find segfault issue - needs fixing

+++ engines/cpp/src/lexer.cpp
@@ -163,7 +163,7 @@ std::ostream& operator<<(std::ostream& out, const Token& tok) {
 // const std::regex NON_STRING_SPACE_REGEX
 
 Lexer::Lexer(void) {
-  assert(tokenRegexMap.size() == TokenType::NUM_TOKEN_TYPES && "Must define regex for every token type!");
+  assert(tokenRegexMap.size() == static_cast<uint64_t>(TokenType::NUM_TOKEN_TYPES) && "Must define regex for every token type!");
 }
 
 bool Lexer::validateOptionsAndSource(std::string sourceCode) {
